requirements-full.txt
# Core API dependencies
fastapi>=0.116.1
uvicorn>=0.35.0
python-dotenv>=1.1.0
pydantic>=2.11.5

# AI and processing - Gemini
google-genai>=1.23.0

# LangChain for LLM formatting
langchain>=0.3.25
langchain-core>=0.3.64
langchain-community>=0.3.24
langchain-google-genai>=2.1.5
langchain-openai>=0.3.21
langgraph>=0.4.8
langgraph-prebuilt>=0.2.2
langsmith>=0.3.45

# AI tools
openai>=1.84.0
tavily-python>=0.7.5
tiktoken>=0.9.0

# YouTube processing - using yt-dlp for better reliability
yt-dlp>=2025.7.21

# Transcription service
fal-client>=0.7.0

# Chinese text processing
opencc-python-reimplemented>=0.1.7

# HTTP requests
httpx>=0.28.1
requests>=2.32.3

# Chromium automation - using pyppeteer (correct version)
pyppeteer>=0.0.25

# Utilities
numpy>=2.2.6
tqdm>=4.67.1
rich>=14.0.0
more-itertools>=10.7.0

# Audio processing
pydub>=0.25.1

# Alternative YouTube downloader
pytubefix>=9.4.1

# AI agents
smolagents>=1.19.0